Fix some problems reading in NCBI databases on a PPC (big-endian) machine. The db version number and db type were not being converted to the correct endianness so the open was failing.  Also fixed some problems when building with machines that have older version of endian.h.  Also fixed some warning when compiling on a 32-bit machine.
Fix a typo in esl_sqio.c that prevented the second example from compiling

Removed an unused define (__bswap32) from esl_config.h.



git-svn-id: https://svn.janelia.org/eddylab/eddys/easel/trunk@483 db3e70e4-12e1-0310-b23f-81a6e3d09ba7

+++ esl_sqio_ncbi.h
@@ -34,8 +34,8 @@ typedef struct esl_sqncbi_s {
   uint64_t   total_res;            /* total number of residues                 */
   uint32_t   max_seq;              /* longest sequence in the database         */
 
-  off_t      hdr_off;              /* disk offset in .pin to header index      */
-  off_t      seq_off;              /* disk offset to .pin to sequence index    */
+  uint32_t   hdr_off;              /* disk offset in .pin to header index      */
+  uint32_t   seq_off;              /* disk offset to .pin to sequence index    */
   
   int        index;                /* current sequence index in the database   */
